+++ packages/runtime/src/account.ts
@@ -395,6 +395,8 @@ class Asset {
     this.id = assetId;
     const base: BaseModel = new BaseModelI();
     this.definitions = {
+      ...base,
+      name: assetName,
       creator: creator,
       total: BigInt(def.total),
       decimals: def.decimals,
@@ -407,8 +409,7 @@ class Asset {
       manager: def.manager,
       reserve: def.reserve,
       freeze: def.freeze,
-      clawback: def.clawback,
-      ...base
+      clawback: def.clawback
     };
   }
 }